+++ BinanceApiAdapter/Types/Balance.cs
@@ -9,7 +9,7 @@ namespace CryptoSwapMaster.BinanceApiAdapter.Types
     public class Balance
     {
         public string Asset { get; set; }
-        public double Free { get; set; }
-        public double Locked { get; set; }
+        public decimal Free { get; set; }
+        public decimal Locked { get; set; }
     }
 }
