+++ BinanceApiAdapter/Types/Filter.cs
@@ -5,17 +5,17 @@ namespace CryptoSwapMaster.BinanceApiAdapter.Types
     public class Filter
     {
         public FilterType FilterType { get; set; }
-        public double MinPrice { get; set; }
-        public double MaxPrice { get; set; }
-        public double TickSize { get; set; }
-        public double MinQty { get; set; }
-        public double MaxQty { get; set; }
-        public double StepSize { get; set; }
-        public double MinNotional { get; set; }
+        public decimal MinPrice { get; set; }
+        public decimal MaxPrice { get; set; }
+        public decimal TickSize { get; set; }
+        public decimal MinQty { get; set; }
+        public decimal MaxQty { get; set; }
+        public decimal StepSize { get; set; }
+        public decimal MinNotional { get; set; }
         public int Limit { get; set; }
         public int MaxNumAlgoOrders { get; set; }
-        public double MultiplierUp { get; set; }
-        public double MultiplierDown { get; set; }
-        public double AvgPriceMins { get; set; }
+        public decimal MultiplierUp { get; set; }
+        public decimal MultiplierDown { get; set; }
+        public decimal AvgPriceMins { get; set; }
     }
 }
